backend/callback-request/requirements.txt
pydantic>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
